codegen: X86 64-Bit now longer supports pushl. Its dead. They killed it. :(
This commit replaces all of our pushl instances with pushq and uses the r
registers. Same for pop.

This may break your codegen logic so please review your parts of the language.
With this commit gcc only complains about a missing __cbp_main symbol in a
simple test case. We need to fix our main0 label to match the __cbp_main call.

git-svn-id: https://elmar.ips.cs.tu-bs.de/svn/cbp_ws1011@182 a60ba226-48e3-df11-83e3-000476a39db3