No main function in assembler. It is renamed to _cbp_main and called by the main function in CBP_main.c. (CBP_main4SDL.c when using SDL library.) You should compile CBP_main(4SDL).c and pass it to the linker when linking a cbp generated program. Otherwise the linker won't like it.
The compiler now accepts a 2nd argument. If you want underscores in front of functions pass "underscore" or "nounderscore" for no underscores. Simple, isn't it? If no 2nd argument is passed default is: APPLE -> underscore, Unix -> no underscore, Windows -> no underscore. Maybe, this is unnecessary if there is a unique default for Windows.

git-svn-id: https://elmar.ips.cs.tu-bs.de/svn/cbp_ws1011@203 a60ba226-48e3-df11-83e3-000476a39db3