Improved readability further
The commas for commands are now generated automatically.
This lead to automatic generation of addresses through
a small embedded language.

Allowed are addresses of the form "base + displacement + offset * multiplier"
or derivations of this which are allowed in assembler
(see Mnemonics.h).

The syntax for commands is something like the following:
Command("mov")("%ebp")("%esp") or Command("mov")(Reg("esp") + offset_it)("%eax").
The first is the mnemonic and then arguments may follow.
Allowed arguments are: registers, numeric constants and addresses.


git-svn-id: https://elmar.ips.cs.tu-bs.de/svn/cbp_ws1011@208 a60ba226-48e3-df11-83e3-000476a39db3